pyproject.toml
[tool.poetry]
name = "drf-recipe-04"
version = "0.1.0"
description = ""
authors = ["ready-oun <heeyn.lim@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
django-environ = "^0.11.2"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
psycopg2 = "^2.9.9"
pillow = "^10.3.0"
gunicorn = "^22.0.0"
requests = "^2.31.0"
boto3 = "^1.34.98"
channels = "^4.1.0"
daphne = "^4.1.2"
channels-redis = "^4.2.0"
black = "^24.4.2"
django-ckeditor = "^6.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
